+++ src/main/java/org/akcjamis/webapp/repository/ChristmasPackageRepository.java
@@ -2,8 +2,6 @@ package org.akcjamis.webapp.repository;
 
 import org.akcjamis.webapp.domain.ChristmasPackage;
 
-import org.json.JSONArray;
-import org.json.JSONObject;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.*;
@@ -14,7 +12,7 @@ import java.util.List;
 /**
  * Spring Data JPA repository for the ChristmasPackage entity.
  */
-public interface ChristmasPackageRepository extends JpaRepository<ChristmasPackage,Long> {
+public interface ChristmasPackageRepository extends JpaRepository<ChristmasPackage, Integer> {
 
     Page<ChristmasPackage> findByEvent_year(Short year, Pageable var1);
 
@@ -23,9 +21,9 @@ public interface ChristmasPackageRepository extends JpaRepository<ChristmasPacka
         "left join fetch subpackages.subpackageNotes subpackageNotes " +
         "left join fetch christmasPackage.christmasPackageNotes christmasPackageNotes " +
         "where christmasPackage.event.year =:eventYear and christmasPackage.id =:id")
-    ChristmasPackage findByEvent_yearAndId(@Param("eventYear") Short eventYear, @Param("id") Long id);
+    ChristmasPackage findByEvent_yearAndId(@Param("eventYear") Short eventYear, @Param("id") Integer id);
 
-    ChristmasPackage findById(Long id);
+    ChristmasPackage findById(Integer id);
 
     @Query(value = "select christmasPackage from ChristmasPackage christmasPackage " +
         "left join fetch christmasPackage.subpackages subpackages " +
@@ -36,7 +34,7 @@ public interface ChristmasPackageRepository extends JpaRepository<ChristmasPacka
             "where christmasPackage.event.year =:eventYear")
     Page<ChristmasPackage> getList(@Param("eventYear") Short eventYear, Pageable page);
 
-    List<ChristmasPackage> findByFamily_id(Long id);
+    List<ChristmasPackage> findByFamily_id(Integer id);
 
-    ChristmasPackage findByFamily_idAndEvent_year(Long id, Short year);
+    ChristmasPackage findByFamily_idAndEvent_year(Integer id, Short year);
 }
