+++ src/main/java/org/akcjamis/webapp/repository/search/ChristmasPackageNoteSearchRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.elasticsearch.repository.ElasticsearchRepository
 /**
  * Spring Data ElasticSearch repository for the ChristmasPackageNote entity.
  */
-public interface ChristmasPackageNoteSearchRepository extends ElasticsearchRepository<ChristmasPackageNote, Long> {
+public interface ChristmasPackageNoteSearchRepository extends ElasticsearchRepository<ChristmasPackageNote,Integer> {
 }
