+++ src/main/java/org/akcjamis/webapp/repository/search/ContactSearchRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.elasticsearch.repository.ElasticsearchRepository
 /**
  * Spring Data ElasticSearch repository for the Contact entity.
  */
-public interface ContactSearchRepository extends ElasticsearchRepository<Contact, Long> {
+public interface ContactSearchRepository extends ElasticsearchRepository<Contact,Integer> {
 }
